Rotate: create rotated copy using temporary directory name during copy
While creating rotated copy, create rotated copy with prefix ".tmp"
first and rename to final rotated copy only after the copy is complete.

This makes it obvious which rotated copies are partial because in case
the system is shutdown or power is lost while copy rotation is in progress
the rotated copy has incorrect name and it's obvious that files may be
missing from that copy.

Note that this patch doesn't implement logic to automatically remove
such partial copies instead of full copies.