Rotate: add timestamp file to keep track of last rotate
Avoid doing rotates too often if the backup directory happens to have
recent files or directories. Add explicit timestamp file for the last
rotate.

Note that we update the rotate timestamp at the start of the rotate and
and the end of rotate to avoid doing it too often if it's really slow.