Workaround Lintian braindamage requesting non-systemd support
We only support systemd and we have dependency for systemd but lintian
still complains about not supporting systems without systemd.

Just ignore this heuristics. Unfortunately, this requires adding a
new crap file to the installed system because this heuristic is considered
as being in binary state, not in the building state.