rsync: Add flag --delete-excluded to re-sync
When backup already contains files in folders that later get excluded by
the configuration changes, the rsync process starts to exit with status
code 23 because it cannot match the directory hierarchy because the
directories which contain now-excluded but historically allowed files
cannot be cleaned up in the backup.

Adding --delete-excluded should be enough to give rsync the permission
to delete the files that would have never been backed up if the current
exclude-configuration had been in use from the start.

This naturally doesn't change the old versioned snapshots, just the latest
version and the snapshots that will be created from it in the future.